src/requirements.txt
clickhouse-driver==0.2.3
paho-mqtt==1.6.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
tzdata==2022.1
tzlocal==4.2
